+++ src/main/java/MainWindow.java
@@ -1,13 +1,11 @@
-import Duke.WagwanException;
-import Duke.WagwanUi;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.ScrollPane;
 import javafx.scene.control.TextField;
 import javafx.scene.image.Image;
-import javafx.scene.layout.*;
-
-import java.io.IOException;
+import javafx.scene.layout.AnchorPane;
+import javafx.scene.layout.VBox;
+import wagwan.WagwanUi;
 
 /**
  * Controller for MainWindow. Provides the layout for the other controls.
@@ -27,6 +25,9 @@ public class MainWindow extends AnchorPane {
     private Image userImage = new Image(this.getClass().getResourceAsStream("/images/mike.jpeg"));
     private Image wagwanImage = new Image(this.getClass().getResourceAsStream("/images/KSI.jpeg"));
 
+    /**
+     * Initialises Wagwan GUI.
+     */
     @FXML
     public void initialize() {
         scrollPane.vvalueProperty().bind(dialogContainer.heightProperty());
@@ -45,7 +46,7 @@ public class MainWindow extends AnchorPane {
      * the dialog container. Clears the user input after processing.
      */
     @FXML
-    private void handleUserInput() throws WagwanException, IOException {
+    private void handleUserInput() {
         String input = userInput.getText();
         String response = wagwan.getResponse(input);
         dialogContainer.getChildren().addAll(
