+++ src/main/java/command/MarkCommand.java
@@ -1,18 +1,25 @@
-package Command;
-import Duke.WagwanException;
-import Duke.WagwanUi;
-import Duke.Storage;
-import Duke.TaskList;
+package command;
+
 import java.io.IOException;
 
+import wagwan.Storage;
+import wagwan.TaskList;
+import wagwan.WagwanException;
+import wagwan.WagwanUi;
+
 /**
-* Command that deletes a task from the TaskList when executed.
+* Command that marks a specified task as complete when executed.
 */
-public class DeleteCommand extends Command {
-    boolean isExit;
+public class MarkCommand extends Command {
+
+    private boolean isExit;
     private String userAction;
 
-    public DeleteCommand(String userAction) {
+    /**
+     * Constructor for a find command.
+     * @param userAction the index of the task to be marked as completed.
+     */
+    public MarkCommand(String userAction) {
         this.isExit = false;
         this.userAction = userAction;
     }
@@ -21,9 +28,9 @@ public class DeleteCommand extends Command {
     public String execute(TaskList tasks, WagwanUi ui, Storage storage) throws WagwanException {
         try {
             int index = Integer.parseInt(this.userAction);
-            String deleteTaskMessage = tasks.deleteTask(index);
+            String markTaskAsDoneMessage = tasks.markTaskAsDone(index);
             storage.save();
-            return deleteTaskMessage;
+            return markTaskAsDoneMessage;
         } catch (IOException e1) {
             throw new WagwanException(e1.getMessage());
         } catch (NumberFormatException e2) {
@@ -37,6 +44,6 @@ public class DeleteCommand extends Command {
 
     @Override
     public String toString() {
-        return "this is a delete command : delete " + userAction;
+        return "this is a mark command : mark " + userAction;
     }
 }
