+++ src/test/java/wagwan/DeadlineFormatTest.java
@@ -1,9 +1,11 @@
-package Duke;
-import org.junit.jupiter.api.Test;
+package wagwan;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.time.format.DateTimeParseException;
 import java.util.ArrayList;
-import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import org.junit.jupiter.api.Test;
 
 public class DeadlineFormatTest {
