+++ src/test/java/wagwan/TaskListTest.java
@@ -1,7 +1,10 @@
-package Duke;
-import org.junit.jupiter.api.Test;
-import java.util.ArrayList;
+package wagwan;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import java.util.ArrayList;
+
+import org.junit.jupiter.api.Test;
 
 public class TaskListTest {
